api/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
asyncpg==0.30.0
redis==5.0.1
python-dotenv==1.0.1
python-dateutil==2.9.0
